+++ src/test/live-coding-playground/src/test/java/quickimport/QuickImportCollections.java
@@ -1,0 +1,7 @@
+package quickimport;
+
+public class QuickImportCollections {
+    List<String> list;
+    Set<String> set;
+    Map<String,String> map;
+}