+++ src/test/live-coding-playground/src/main/java/lombok/slf4j/ReplaceWithSlf4j.java
@@ -1,0 +1,18 @@
+package lombok.slf4j;
+
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class ReplaceWithSlf4j {
+
+    private static final Logger log = LoggerFactory.getLogger(ReplaceWithSlf4j.class);
+
+    private String name3;
+
+    public static void main(String[] args) {
+        log.debug("");
+    }
+
+
+}