+++ src/main/kotlin/com/github/victorrentea/livecoding/extracthints/SyntaxExtractableSectionsVisitor.kt
@@ -1,10 +1,12 @@
 package com.github.victorrentea.livecoding.extracthints
 
 import com.github.victorrentea.livecoding.startLineNumber
+import com.intellij.lang.ASTNode
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.psi.*
 import com.intellij.psi.util.PsiTreeUtil
 import com.intellij.refactoring.suggested.startOffset
+import org.mozilla.javascript.ast.IfStatement
 
 class SyntaxExtractableSectionsVisitor {
     companion object {
@@ -24,9 +26,8 @@ class SyntaxExtractableSectionsVisitor {
 
                     if (tentativeStatements.any { hasOrphanBreakContinueChildren(it)}) continue
 
-                    if (tentativeStatements.any {
-                            PsiTreeUtil.findChildrenOfType(it, PsiReturnStatement::class.java).isNotEmpty()
-                        }) continue // TODO remove:  In fact, we could extract any block if all the exec paths end with a RETURN or THROW
+                    if (partialReturns(tentativeStatements)) continue
+                    
 
                     sections += tentativeStatements
                 }
@@ -35,6 +36,36 @@ class SyntaxExtractableSectionsVisitor {
 
         element.children.forEach { visitElement(it) }
     }
+
+    private fun partialReturns(tentativeStatements: List<PsiStatement>): Boolean {
+        // returns before the last statement
+        if (tentativeStatements.subList(0, tentativeStatements.size - 1).any { containsReturns(it) })
+            return true
+
+        val lastStatement = tentativeStatements.last()
+        if (!containsReturns(lastStatement)) return false
+
+        return !returnsOnAllBranches(lastStatement)
+    }
+    private fun returnsOnAllBranches(statement: PsiStatement):Boolean {
+        if (statement is PsiReturnStatement) return true
+        if (statement is PsiBlockStatement) return !partialReturns(statement.codeBlock.statements.toList())
+        if (statement is PsiIfStatement) {
+            if (statement.thenBranch?.let{returnsOnAllBranches(it)} == false) return false
+            if (statement.elseBranch?.let{returnsOnAllBranches(it)} == false) return false
+            return true
+        }
+        if (statement is PsiLoopStatement) {
+            return statement.body?.let {containsReturns(it) } == false
+        }
+        return true;
+    }
+
+    private fun containsReturns(it: PsiStatement) =
+        PsiTreeUtil.findChildrenOfType(it, PsiReturnStatement::class.java).isNotEmpty()
+
+    private fun incompleteReturn(it: PsiStatement) =
+        PsiTreeUtil.findChildrenOfType(it, PsiReturnStatement::class.java).isNotEmpty()
 
     private fun hasOrphanBreakContinueChildren(statement: PsiStatement): Boolean {
         // a break or continue must all have a parent for/while/dowhile that is a child of statement