+++ src/test/live-coding-playground/src/main/java/playground/trivia/Game.java
@@ -133,10 +133,7 @@ public class Game {
             if (currentPlayer == players.size()) currentPlayer = 0;
             return true;
          }
-
-
       } else {
-
          System.out.println("Answer was corrent!!!!");
          purses[currentPlayer]++;
          System.out.println(players.get(currentPlayer)