+++ CHANGELOG.md
@@ -53,6 +53,7 @@
 - made the method complexity renderer more silent
 - temporary disable introduce guard inspection
 ## 0.0.15
-- Improved Replace with @RequiredArgsConstructor to use syntax references not just parameter and field names
+- Improved Replace with @RequiredArgsConstructor to use syntax references not just parameter and field names; Also made available the inspection if INFORMATION level on all constructor text
 - Stopped Add @RequiredArgsConstructor if there is a constructor in the class
 - Added a few more static imports
+- Made quick import add imports for Set,List and Map